Updates schema.rb, and adds schema.test.rb.
We switched to MySQL for development and also pulled in a new migration that had
been added while still using SQLite3... so need to re-run migrations using MySQL
to get an up-to-date schema.rb.

Meanwhile, we use explicitly set the env var SCHEMA to 'db/schema.test.rb'
in environments/test.rb, so that when RAILS_ENV env var is 'test' then this
schema will be used when loading/dumping the schema.

Also,
* Removes problematic integration spec that uses a roundabout way of testing
  user permissions. For some reason the page intermittently takes a long time
  (> 120 sec) to load on Travis, which causes intermittent failures. Can't
  replicate on development, so can't get to the bottom of it. Been bugging us
  for weeks (pun intended!). If we want to test perms let's find another way
  that doesn't rely on iffy Capybara methods.