Creates additional BatchItems for Asset children
Before this change, a BatchItem only represented a top-level Asset record, even
though the Asset record may have children (e.g. DigitalInstantiations) that are
ingested by running concurrent jobs.

The problem with this is that an Asset children might fail on ingest without
that failure being properly reflected in any BatchItem that is part of the
Batch.

This change creates new BatchItem objects for each Asset child, so that the
status of those child objects can be recorded accurately in their own BatchItem
instances.

Also...
* Adds very opinionated script for calculating batch ingest duration from Rails
  and Sidekiq logs. Used as a command line tool to compensate for shortcomings
  in current reporting of ingest performance.
* Fixes Faker deprecation warnings.