Use stateless ID minter for new AAPB ID
We were running into issues in deployed environments with our stateful minter,
which requires providing concurrent access to a single persistence layer where
the minter state is stored. Our IDs are random, and we don't have a use case for
"replaying" the minting of them, so we can switch to a stateless minter and
avoid the persistence layer requirement, and associated headaches.

Our stateless minter replaces NOID, and is just the first 12 chars of a UUID
with the constant prefix "cpb-aacip-" prepended to it.

For every ID minted, we do a lookup in Solr to see if the ID exists yet, and if
it does, we regenerate a new random ID.

Also...
* Provides an accessor to AMS.seeds, for manually creating seed data. This has
  proen useful more than once in development, so added it in.