Raise errors when saving model fails
Previously, validation errors, and potentiall other errors that would cause the
actor stack's #create method to return false, were not bubbling up to the
rescue_from clause of the batch item ingest job, and thus were getting
overlooked as the batch item went on to report a 'completed' status.

Now, if the actor stack fails, it will raise an exception that will be reported
on the batch item. If it's a validation error, then that message will show up,
otherwise, it will be a generic message, which can become more detailed as we
encounter errors that we want to more accurately report.

Fixes #292.

Also...
* In ZippedPBCoreDigitalInstantiationMapper...
  * removes duplicate method definition for #digital_instantiation_attributes.
  * replaces #get_row_data with memoized #row_data method.
* Modifies test for 'AAPB PBCore - Zipped' ingest for better output on failures.