Fix bug with cascade delete
The previous attempt didn't work for models created with actors because deleting
the Asset made it impossible to access the Asset#ordered_members (although for
reasons unknown, this wasn't a problem in the Asset model spec).

This change deletes objects from Asset#members rather than Asset#ordered_members
which seems to work.

The point of adding cascade delete, was so that we can identify Assets that were
ingested via the 'AAPB PBCore XML Zipped' batch ingest, but which had child
objects that failed ingest. In this case, we want to delete the Asset and all
it's ingested children, so that we can re-ingest it with the same IDs.

This PR also adds a utility, AMS::AAPBPBCoreZippedBatchIngestFixer, which does
just that, when given a batch id.