Do not switch status to expunged
Per request of CJ Colvard of IU, we don't want the Batch Ingest gem to have an
additional status of 'expunged'. Instead, we'll keep the status as-is, and each
batch item will instead recognize when it's repo_object_id has been deleted, and
report the associated repo object as 'Gone', whether or not the repo object was
explicitly expunged, or removed via some other process.

This also refactors AAPBPBCoreZippedBatchIngestFailureFinder and
AAPBPBCoreZippedBatchIngestExpunger into a single object called
BatchIngestUtility, the idea being that the BatchIngestUtility operates
on a batch, and performs some idempotent maintenance duties.

Updates to latest version of hyrax-batch_ingest gem.