Fixes broken Travis builds
We were getting errors that the version of ChromeDriver was incompatible with
the version of Chrome that Travis was running.

After much debugging, we found that a recently released version of Chrome was
not compatible with the 'trusty' Linux distribution that our Travis CI config
was specifying.

We upgraded our Linux distribution from `trusty` to `xenial`, but in so doing,
our Travis configuration for JDK `oraclejdk8` became incompatible.

Upgrading the distribution from `trusty` to `xenial` then caused an issue with
running `oraclejdk8` (required by Fedora and Solr), because `xenial` is
incompatible with `oraclejdk8`.

We changed our JDK config to use `openjdk11` and changed how we start xvfb, from
calling an init.d script, to using Travis config `services` option. This seems
to have caused a lot of Java errors as Travis was booting up the test suite, but
the tests ran to completion and passed.

The Java errors should be investigated further, but should not block this PR as
we need to get builds passing again quicly.

We also added `--fail-fast` and `format --docuemntation` options to `.rspec`.
Going to leave these in for now, but may want to remove them later if they
prove to be more annoying than useful.