Conditional cascade delete of failed records
When a record fails during ingest, our policy is to delete all other records
that are part of that 'family', so that a subsequent ingest can re-add the
whole set of records.

However, there is a use case for adding new child records, such as
DigitalInstantiations, to an existing Asset that was successfully ingested
previously. If that new child record fails, we were erroneously deleting
the whole family of records (i.e. the parent Asset and it's children)

This PR fixes that by making sure that, if we are deleting record that failed
the ingest, then we check to see which batch it is part of, and do NOT delete
the rest of the record family if the failed object is part of a new ingest.