Refactors permission configuration
Hyrax's authorization system has a few moving parts.
1. The applications's `Ability` class, which includes `CanCan::Ability`.
2. `Hyrax::Ability` and `Hydra::Ability` mixings that are injected into the
    app's `Ability` class.
3. Application data stored in DB tables that inform permissions defined in
   the `Hyrax::Ability` and `Hydra::Ability` mixins.

Previously our permissions defaulted to be overly permissive, and we wanted
to switch that to be restrictive by default, and selectively permissive based
on group.

Also, some of our custom permissions in the app's `Ability` class were being
overwritten by logic within the `hyrax` and `hydra-access-controls` gem.

As a result, some of our tests were passing based on false positives due to the
overly permissive rules, or to rules we had specified that were not having an
effect.

* Adds new rules to `Ability` class and refactors how they are applied by not
  relying on the `custom_permission` method that is provided by the
  `Hydra::Ability` class from the `hydra-access-controls` gem.
* Adds helper methods to `PBCoreXMLItemIngester` class to help determine
  permissions related to ingest of PBCore XML.
* Renames role from "user" to "test-group" when creating `user_with_role`
  variables in some specs. This is a bit more clear that the role is arbitrary.