2038 Annotations (#457)
Creates `Annotation` model for managing arbitrary PBCore annotations.

This arose out of the need for `AdminData` records to be more versatile. The limitations were that there was a one-to-one association between `Asset` records in Fedora and `AdminData` records in the db, and the fields of the `AdminData` model basically represented specific annotation elements within the PBCore.

What we needed was a was for users to be able to add multiple new PBCore annotations to a single Asset record without requiring a database schema update.

This PR covers adding the `Annotation` model, integrating it into the existing data model, integrating into the actor stack for CRUD operations, and adding a migration script to populate the `annotations` table with data from existing recrods.