Fixes NoMethodError in export records notification
This was ironically causing the job to fail on the "success" notification email,
which then was not being rescued from. But a recent change adds a rescue from
the job and the false negative was exposed.

This is a super quick fix for the immediate problem. More proper of the export
system is needed though.

Fixes #490.