Fixes issue with queries of longs ID lists
ID lists in some cases are very long, too long for Solr to handle. There was
prevously logic to break those lists up into multiple queries and recombine the
the results. This was done in 2 locations, but missing for Instantiation exports
which were probably broken, but there wasn't a test for it.

This PR does a few things to remedy that;
* Change Search::Base to have a default behavior of querying with a
  Blacklight::Solr::Repository instance, without CatalogController config.
* Then, Search::IDSearch extends Search::Base.
* Then, Search::CatalogSearch (replaces AssetsSearch) extends Search::Base, but
  *does* bring in the config from CatalogController (including it's
  SearchBuilder and all of the processor chain associated therewith).

Also adds PushToAAPBJob which is used by PushesController instead of
ExportRecordsJob.