Updates CodeDeploy access key
Continuous deployment is achieved by telling Travis to deploy using AWS CodeDeploy
after tests pass. This requires use of a valid API access key for the IAM user that
is used to do the deployment. That key just changed, so our config needs to change
accordingly, and to test the automatic deployment, we temporarily change the target
of the deployment to be the CodeDeploy application and deployment group that are
tied to our demo stack (i.e. not our production stack).

After we are satisfied continuous deployment is working for demo stack, we can switch
the CodeDeploy application and deploymenet group variables back to those that
correspond with our production stack.

* Updates access_key_id to match new API key credentials.
* Updates secret_access_key to match new API key credentials. This is encrypted
  using `travis encrypt [secret_access_key_value]` which is part of the `travis` gem,
  which only needs to be run locally, and not in the Gemfile.
* Temporarily sets the CodeDeploy application and deployment group to be those used
  for our demo stack, with the expectation that we will change it back to the
  application and deployment group that are used for production once we're satisfied
  that it works.

Fixes #540.