Fixes deployment info page
The problem was that the script trying to fetch deployment info was using AWS credentials
for the user who has permissions on S3, but not CodeDeploy.

This uses a different set of credentials which it expects to be correctly set in the
environment. If the expected env vars are empty, the error message will say say.