Refactors error handling in MediaController#show
Some error cases were not being covered:
* when SonyCiApi gem returns a SonyCiApi error.
* when the SolrDocument is missing the sonyci_id_ssim field.
* when the sonyci_id_ssim field is there, but empty.

Some of these errors were resulting in ugly 500 errors with HTML rendering. But this controller
is specific to redirect to a Sony Ci URL where the media is, and return the byte stream of the
media, so HTML reponses are out.

Here we always return the redirect on success, and a head response (no body) with the appropriate
HTTP status on failure.

closes #609.