Fix 'Find media in Sony Ci...' button
This button (as a <a> tag currently) is rendered on the Asset Edit interface in the Admin Data
section of the form, and finds Sony Ci media whose filename matches an Asset's GUID. But it
wasn't working on initial page. The behavior was being added in the turbolinks:load event,
which should work (we're doing it elsewhere), but on this page it simply would not. So I
changed it to use the jQuery(document).ready() method, which typically does NOT work because
of turbolinks, but in this case, it works better. But still not sure why.

Closes #608.