Re-indexes Asset after webhook from Sony Ci
The Asset was not being reindexed, and so it was saving on the Asset's admin data, but not the Solr doc.

This explicitly calls Asset#save! to re-index, but Asset#save! also persists in Fedora, which is slow, so if
there's a faster way to just save AdminData and re-index without Fedora, we should do that

Closes #620.