Change FCREPO Urls for New Records
AMS records all have ids prefixed with cpb-aacip- This meeans that
they all end up under the same subtree in Fcrepo, which is causing
performance issues.

For new records, use the last segments of the ID instead of the first
to get a more even distrubution. But still be able to find the existing
records until we can get them all moved.