don't null AdminData value unless attr is present
Before, if an attribute's key and value were completely missing, it would null out the value. I don't believe this is the desired behavior.

I interpret the `!env.attributes[key].present?` to mean "if an attribute **is present** and its value is missing, null out the data".

However, in reality, what this actually does is "if an attribute is present and its value is missing OR the attribute is missing entirely, null out the data", which feels too destructive