✅ Fix the way #query_solr for Hyrax 4.0
This commit will adjust how `#query_solr` works for the upgrade to Hyrax
4.0 which uses Blacklight 7.  The `#search_results` method no longer
takes an argument, so we need to adjust how we call it.  I added the
same method as from Hyrax to call the `Hyrax::SearchService`.  I
couldn't call it directly from Hyrax because (at least in the tests) I
was not able to pass the params because they were not permitted.

Ref
  - https://github.com/samvera/hyrax/blob/4.x-stable/app/controllers/concerns/hyrax/controller.rb#L48-L51