🧹 Wrap recent changes in conditional
This commit wraps the `App.rails_5_1?` conditional around the recent
changes just in case it breaks anything for production app.  The tests
indicate it would not however we don't know how comprehensive the tests
are at this moment.  For `_index_list_default.html.erb`, I decided it
would be easier to just wrap the entire file in the conditional rather
than the specific parts.  This would make it easier to revert as well.