🧹 Make copy of `views` folder
This commit will make a copy of the `views` for the Rails 6.1 upgrade.
This strategy should be relatively safe because few (if any) other devs
will be working on the views while I'm working on the upgrade while
still allowing me to push and deploy to staging and production without
the views being broken.  This way we don't have to set a conditional at
the partial level and makes it easy to bring over the changes after the
upgrade.