🧹 Modify `_index_list_default` partial
This commit will modify the `_index_list_default` partial to use the
default Hyrax 4.0 version.  I was not able to see how a Collection could
be displayed in the default list view, so I essentially removed it by
using the default Hyrax 4.0 version behind the conditional.  If no
changes in behavior are noticed, we can remove the entire partial from
the app so we can use the default Hyrax 4.0 version.