🧹 Fixed the look of `_search_form` partial
This commit will reconcile the differences between the override and the
Hyrax 4.0 version of the `_search_form` partial.  Note that the `All`
and `More options` buttons would be using `btn btn-secondary` classes by
Hyrax 4.0 default, but we are overriding them to use `btn btn-default`
to match the current look of the application.  In the
search_results.css, I removed the search-form form-control height so the
height of the search box would match the rest of the elements.  Now it
has the height of 38px which matches Nurax.