🧹 Modify `_form_progress` partial
This commit will reconcile the overrides made to the `_form_progress`
partial with the Hyrax 4.0 version.  The extra buttons in the
`valid_child_concerns` area were changed from simple form buttons to
`button_tag` elements because the `simple_form` version seems to add a
few extra classes that were not needed.