🧹 Modify `batch_ingest/batches/index and show`
This commit will clean up the `batch_ingest/batches/index` view to use
Bootstrap 4 classes for the table.  Just added a space in the show page
but the other partials its rendering are all in the batch_ingest gem.  I
don't know if they are supposed to look like that or  not but the text
looks a little off to me and might be worth looking into updating.