✅ Fix `AdminAddUserroleAsAdminsetManager` spec
This commit will modify the `AdminAddUserroleAsAdminsetManager` spec to
reflect what's going on in the application.  Before, if you visit the
`admin/admin_set` route it would redirect you to
`dashboard/my/collections` which wouldn't show any admin sets.  Now
we're going directly to `dashboard/collections` which will show all the
admin sets and allow you to edit and do all that stuff.  I think
effectively changing the spec in the manner still tests the
functionality we're looking for.