✅ Fix `CreateAsset` spec
This commit will fix the `CreateAsset` spec by removing the adjusting
the `hyrax/assets/_metadata` partial.  I now understand why the
`collapsed` tag was originally not included in the element.  It was
meant to be automatically opened.  However, it wasn't working as
expected because the div that was supposed to show had the classes
`collapse in` which wasn't working.  We've changed it to `collapse show`
and now it works as expected.