✅ Fix `pushes_spec`
This commit fixes the `pushes_spec` by adjusting the specifics a little
bit. In Hyrax 4/Blacklight 7 the catalog search input has a different
`id` but the `name` remains the same as in previous versions so we'll
use that instead.  I had changed the `Push To AAPB` link to a button
previous so this commit will address that.  Now instead of using
`#click_link` or `#click_button` we'll use `#find` and `#click` to
activate it so it works for both versions of the app.