🧹 Fix #search_fields and #thumbnail_url deprecation
This commit will move the `#search_fields` and `#thumbnail_url` methods
from the Blacklight to the `ApplicationHelper` to avoid the deprecation
warning.

If we ever upgrade to Blacklight 8, the way the facets and the date
range filter might have to be changed.  It is already a lot different in
Blacklight 7 but it still seems to work at least.