:gear: configure sentry gems to 5.4.2 versions
We are noticing weird behavior with bulkrax imports that is related to sentry-rails beings installed. I rolled back to version 5.4.2 to match the version of our other hyrax x bulkrax projects. Additionally, I tried to roll it back to the deprecated sentry-raven gem but that caused conflicts with the other upgrades.