🧹 Add attributes to valkyrie models
This commit will register the new models and also add the attributes.
The approach was to look at the fedora model and extract and conform the
attributes to the new valkyrie models as best as possible.  This is a
first pass and will most likely need to be updated as we go.