give redis time to read AOF during startup
When deployed, the redis pod currently restarts over and over because it
fails the health check before the state can be restored from the AOF
file. The health checks need to be delayed to give time for this to
happen.

The commit corrects the syntax to set the initial delay value.
`healthCheck.initialDelay` is how fcrepo sets its health check delay.