Refactors transaction step for saving admin data
By default, multi-valued input fields display an additional set of input fields that are blank.
These get submitted by the form and are usually discarded, but because we are treating Sony Ci
ID field special, it was erroneously saving additional blank values for Sony Ci ID, which was
resulting in multiple players of the same media dislaying on AAPB when the record was published.

Multiple Sony Ci IDs are allowed in cases of mult-part Assets. They are stored as a serialized JSON
array in the AdminData.sonyci_id field, which is related to Asset reords via a global ID field on
AdminData.gid. And the global ID fields was used instead of a normal foreign key because the
AssetResource used to be just Asset, and stored in Fedora, not the database, and a global ID was
the closest thing to a foreign key across 2 persistence layers that we could think of :)

Fixes #882.