Fix for collections broke batches display
This commit fixes a bug introduced in the fix for collections.

The Hyrax::MyController is shared by both Hyrax::BatchIngest and Hyrax
Collections. To fix collections, we added class to the
load_and_authorize_resource
call in Hyrax::MyController. This broke the batches display in
Hyrax::BatchIngest because batches are not valkyrie resources and cannot
be found in Hyrax's valkyrie cancan adapter.

This commit fixes introduces custom resource authorization for both situations.