Handle contribution object differently
Since we don't want the Contribution objects to show in the view, it
shouldn't be handled via the forms.  Instead, we want it to be handled
as if it was being added through an Asset Resource form.  There's
currently a "Credits" dropdown that now handles this through the
handle_contributors transaction.

By adding the Contribution data to the formatted_row_data before it is
sent off to the make a Hyrax::BatchIngest::BatchItem, we can inject the
information there.