Fix `Contribution` internal resource bug
Given a Contribution that hasn't been migrated to a Valkyrie Resource,
when we look at the contributor attrs, we already have an internal
resource but that needs to be deleted so it can be automatically
generated otherwise we set the string like "#<Class:0x00...>" as the
`internal_resource`.  The `internal_resource` should be "Contribution".