Add rundown (#1055)
Adds new Description Type "rundown" to the `AssetResource` model and all other places that are needed to make that work including the `SolrDocument` model, ingesters (for Batch Ingest), importers (for Bulkrax), controlled vocab config for description type, and to various specs that test the functionality of description types.

Note: this PR also adds the "Rundown" description type to the deprecated `Asset` model for consistency, but the `Asset` model (which was for Fedora) is soon to be removed, along with all of the other Fedora-based models and their supporting code (see issue #1065).